Adelanto Clases
Se crearon las entidades, los dtos, los servicios(están pelados), los repositorios y un package llamado utilities, en el cual hay una clase con la que podremos hacer el mapper de Entidades a DTOS

+++ Taller_Rendimiento/src/main/java/co/edu/unbosque/Taller_Rendimiento/Repository/ClienteRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import co.edu.unbosque.Taller_Rendimiento.Entidades.ClienteEntity;
 
-public interface ClienteRepository extends JpaRepository<ClienteEntity, Long>{
+public interface ClienteRepository extends JpaRepository<ClienteEntity, Integer>{
 
 }
